Implement finding the position and range of a tree
A new compiler flag is required,
    -Yrangepos

Apparently, the best documentation on the flag is the Github
issue in which it was introduced:
https://github.com/scala/bug/issues/2110

Each subtree in a compilation unit knows about its position in the
source file. The problem is, only the starting position is
specified, and there are no means to determine where the part of
sources corresponding to this subtree ends. That's where
`-Yrangepos` appears: it overrides the compiler behaviour to
propagate the end coordinates too.

The flag should be used with caution since it causes instability
https://github.com/scala/bug/issues?q=is%3Aissue%20rangepos%20is%3Aopen
and can disappear at any time, which is indicated by the `-Y`
prefix as opposed to `-X`.

An alternative solution was found here:
https://github.com/ghik/silencer/blob/690719727ee4974bd3a5e3f570be79a40133d3c8/silencer-plugin/src/main/scala/com/github/ghik/silencer/SilencerPlugin.scala#L55-L68
Here they traverse the tree and consider the last known position of
any of the subtrees to be the ending of the initial tree. This
approach works for modules, because it just so happens that a tree
for an invisible function `init()` is created just at the closing
curly bracket. However, the solution fails for methods:

     def hello() = println("Hello, world")
    /\                    /\
    |                     |
    start                 end

There is no subtree generated just after the end of the method's
body, so the last available position is the start of the last
lexeme.

If `-Yrangepos` proves too unreliable, it will be necessary to
adapt the second approach, for example, by finding the closing
pairs of all the opened brackets in the method. Annotations in
other contexts -- in types, in function arguments etc -- also
weren't tested, so maybe there will be problems there too. For now
we'll stick to `-Yrangepos`, so hopefully we won't have to deal
with the ad-hoc method.