Refactor slightly
The method that extracts the annotations satisfying our conditions
from a given tree proved too useful to hide it in another function.
For example, we'll need it for implementing transforms of the
trees since it won't be possible to just change all the trees in
place due to their immutability.