Make tests rebuild when the plugin changes
There was no clear dependency between tests and the compiler plugin,
and so when the plugin was rebuilt, it didn't affect the tests
since their files hadn't been changed. Now, with the idea from
https://github.com/retronym/boxer/blob/00e7deff893afe9cc4e1046aabeca1c4ada44f1a/build.sbt#L31
the /compiler flags/ of the tests have a dependency on the plugin's
jar file. So, now one doesn't have to issue an `sbt clean` and
thus unwillingly remove the unchanged class files of the plugin as
well. Only what needs to be rebuilt gets rebuilt.