Add the jar assembly phase
In order to depend on external libraries, compiler plugins need to
have access to the classfiles from these libraries. Plugins are run
from the compiler and therefore have the same classpath.

Therefore, it makes sense to just package all the dependencies of
the plugin in the same jar file. This way, all the required
classfiles will be within reach of the plugin.

For some alternative approaches and a brief explanation of how
this all works, see https://www.scala-lang.org/old/node/6664