Add an annotation for testing scalameta integration
Create a new annotation that *shouldn't* have any significant
effects, and if it does, that's a sign of a problem. This
annotation reads the annotated code block from the source file,
then parses this block using scalameta, then serializes this block
to a string, then parses the string using the compiler's
facilities, and finally inserts the parsed tree in place of the
original one.

The failure of the tests that use this annotation can indicate
that either the ability of the plugin to correctly specify the
positions of the start and the end of a block have deteriorated
or scalameta doesn't behave as expected.