Create a placeholder for proper method resolution
We need to be able to load methods that aren't in the compiler's
classpath in order to resolve the changes that need to be
introduced in the trees. Currently we still only support the
methods from the compiler's classpath, but it should be possible
to load other classes too by using the macro classloader. It was
difficult for now to make it work, so only a placeholder is
committed.