Implement proper retrieval of annotation functions
Class loaders form an hierarchy, and if a class is already loaded
in a parental class loader, the child loader uses the class loaded
by the parent.

We were using the wrong `analyzer` and ergo the wrong
`findMacroClassLoader`. The function that we called created a
class loader that wasn't a child of the class loader used by the
compiler to resolve the imports in the plugin, and so they loaded
different "versions" of the same scala.meta package.

Now we copy-pasted the code from
scala/src/compiler/scala/tools/reflect/ReflectGlobal.scala
to create a class loader that resolved what is needed in the
sources where macro annotations are described and also uses the
already loaded imports where available, thus re-using the imported
scala.meta.