Change the project structure
The project logically consists of two parts: the annotation trait
that marks the relevant annotations, and the plugin itself. Before,
they were put into the same package, which doesn't make sense given
how they are expected to be used. Now they are separated, and the
main project (which hosted both the marker trait and the plugin)
consists of tests for the plugin.