Fix nested annotations.
It turns out that scalameta's `transform` invalidates the position
info, so we can't apply an annotation function to a subtree and
then still manage to find the annotated parent.

The new approach is really ugly, but it looks like it should work.
So, now we traverse the original tree (as seen by the compiler)
and store the list of annotated subtrees as well as the relevant
annotations for each such subtree.

Then, we reverse the list and traverse it.

Why do we have to reverse the list, you may ask?

Because *I am become the master of kludges*. (Sorry for jokes in
the commit message: it was an unpleasant experience, I need to
vent). So, I thought, if the positions are invalidated, then what
*doesn't* change when we apply an annotation function? The only
stable things that I've found were:

  * The changes are local: they can't yet go outside the annotated
    scope.
  * The names of the annotated subtrees.

The names are really hard to track without typechecking, so I've
decided to do with the spatial locality of changes.

So, if we apply the annotations in the reverse order of
encountering them in the source file, then the order of the trees
that will be processed later as seen by the `transform` method
doesn't change.

For example, in the tree

    import x._

    @AnnotationA
    class A {
        @AnnotationB
        class B
    }

    @AnnotationC
    class C { }

if we first apply `AnnotationC`, then both `AnnotationA` and
`AnnotationB` are going to have the same positions as they did
before if we look at the order of traversal performed by
`transform`. The changes begin with `AnnotationC` and onward.

So, we begin with the end of the file and proceed to the
beginning, applying the annotation functions.

This approach may still be viable later, when we deal with
companion objects properly, if we do it in a lazy fashion.