Fix a bug where everything was considered an annottee
The code by mistake thought that every definition was annotated
by our annotations. It didn't cause any significant problems
because the actual list of annotations to be applied was empty for
such false positives. Nonetheless, it did cause parsing with
scalameta to be lazy no longer, since the code was parsed in any
case if there was at least one definition in the source file.