Avoid custom namers and a lot of bugs.
The custom namer that was used to avoid crashes wasn't sufficient
for the task, as shown by crashes that happen when a namer enters
a symbol into a package that is already present in the root
context, for example, the empty package.

So, in order to avoid namers being such a pain, we'd need to
modify the namer further to avoid entering any symbols at all,
since any of them could theoretically still make their way into
the global context.

The new workaround seems more cludgy but at the same time robust in
a way. We avoid custom namers altogether and run the effectful
namer provided by the compiler--but first we put everything into a
new package with the name `'paradise-ng`, hoping that no one else
ever tries to access the package.

Now namers do add trash to the global context, but this trash isn't
in conflict with the proper definitions since now they're in a
separate namespace.