Remove the fake package from the context
During typing, we put the temporary tree into a fake package so
that it has no chance of interfering with the real definitions.
However, the solution was only partial because

file A:

    class A { }

file B:

    val a = new A

Before, when we were done with file A, there were two definitions
of `A` in the context: the real one, in the package `<empty>`, and
the fake one, `'paradise-ng.A`. But when we went on to deal with
file B, we put it, too, into the fake `'paradise-ng`. So when we
tried to assign types to the temporary tree for file B, the symbol
assigned to `A` was, naturally, `'paradise-ng.A`, even though that
symbol shouldn't exist already.

Now, this problem could be mitigated in several different ways,
for example, by randomizing the fake package name, but it looks
like nothing can break if we simply remove all mentions of the
fake package after we're done.