Partially implement REPL integration
Now one can define macro annotations in the RELP and apply them
at the next line (not on the same one though).

The only caveat is that one can't apply the macros that change the
set of top-level definitions. For example,

    class FooAnnotation extends ParadiseNgAnnotation {
        def append(tree: Tree): Tree = {
            q""" class Foo { def bar() = 34 }"""
        }
    }

will only work if it's applied to class Foo, like this:

    @FooAnnotation class Foo { }

Then running `(new Foo).bar` correctly outputs 34. If one does

    @FooAnnotation class Bar { }

then neither `Foo` nor `Bar` become defined.

This is because of
https://github.com/scalamacros/paradise/issues/19
scala's REPL determines the set of top-level definitions by
examining the tree immediately after the parsing phase, not waiting
for it to be typed.

There is a workaround in the old `scalameta/paradise` project,
in the `ReplIntegration.scala` file, so we need to adapt that to
our case.