Make tests a separate subproject
Now options to scalac are assigned more granularly. The sources
in `tests/src/main/` aren't being tested, instead they help with
the testing, and so the compiler plugin isn't activated for them.
It saves some time during compilation, but more importantly, it
explicitly ensures that annotations aren't magical and don't need
the plugin to compile.