Implement some companion object support
Now when one applies an annotation to a class that has a companion
object, what goes to the function is `Term.Block(tree, companion)`.
Then, if the function returns a block of the form `Term.Block(tree,
companion)`, the `tree` is assigned to the old tree, and
`companion` is assigned to the companion.

Now, it's an error to return a `Term.Block(a, b)` if there is no
companion object defined. A companion object does not get created
all by itself.

Also, there are problems with annotated companion objects if the
class is also annotated. Now everything works like this: if the
class is defined above the companion object, then first an
annotation on the object is invoked, then the result of it is
transformed with the annotation on the class. However, if the
class is defined below the companion object, the annotation on the
companion object can be ignored.