Add support for "companions" of type synonyms
It seems that `scalac` itself does not consider the eponymous
objects of types to be their companions. And it makes sense:
should the companion object get access to private fields of the
original class? Then it's a huge breach of the permission system:
just make a type synonym with a companion, and you can do whatever.
Otherwise, there is no link between a type synonym and its
"companion" and there is no need to connect them in any way.

However, the old `paradise` does treat objects eponymous to
type synonyms the same way it does with clases etc, in order to
allow annotating type synonyms while simultaneously adding new
methods or something. So we do it too.