Small optimisation of meta tree searches
Today, my computer broke. So, I've had to use an old netbook on
which the whole test pass requires 45 minutes. I've spent the
whole day making myself (somewhat) comfortable in the new
environment, so no substantial code today. Just a small
optimisation: we don't go inside the trees that we know we won't
need to find a subtree with a given position. Saves a lot of time.