Simplify class instantiation code
Now, we just try every constructor in order, and if one of them
returns a proper value, we return it.

This has arisen because now we also need to pass as parameters
objects more complex than the primitive values. The code shouldn't
cause problems except in some obscure cases when one set of
parameters can be used for many constructors, and in such cases,
we can't even tell the constructors apart anyway.

Also, it seems that multiple constructors are used rarely in Scala.
We could even refuse to work altogether if more than one
constructor was defined and probably no one would complain, but the
current approach didn't require more effort and covered more cases.