Don't quit after the first error in annotation instantiation
Before, when a single error was found in annotation instantiation,
the whole plugin aborted. Now, all other annotations are checked
first. This way, the end users will see more error messages and be
able to fix more of the mistakes without doing a recompilation to
see each one. It can be a frustrating experience to fix an error
just to learn that there are probably many more hidden.