Preserve error message positions in unmodified subtrees
We change only those parts of the initial tree that we absolutely
need to. Before, we parsed the whole source file anew, losing all
the positions in the process. For example, if in this code

    class Foo { }
    class Bar { }

we add a new method to Foo, then after this commit `Bar` will be at
the same line and offset as before the change. It doesn't
correspond to reality (there are some extremely compact characters
in `Foo`!) but it makes sense when showing the error messages.