Implement intuitive behavior of annotation combination
So, now there is a behavior that at least in theory can be
intuitive.

    @A
    @B
    @MacroC
    @D
    @E
    @MacroF
    @G
    class Z

expands like this: first, `@MacroF` is applied to `@G class Z`,
then to each resulting tree that is not a companion object `@D @E`
are prepended, then `@MacroC` is applied to the result, and
finally, to each non-companion tree of the result `@A @B` is
prepended.